# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: FortiFi API
  description: |
    API for FortiFi paywall hardening service. This API provides endpoints for:
    - Token generation and validation
    - Article content delivery
    - Rate limiting and security
    - User management and subscriptions
    
    ## Authentication
    
    The API uses JWT tokens for authentication. Include the token in the Authorization header:
    
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Rate Limiting
    
    The API implements rate limiting to prevent abuse:
    - 100 requests per hour per IP address
    - 20 requests per minute for demo endpoints
    - Temporary IP blocking for repeated violations
    
    Rate limit information is included in response headers:
    - `X-RateLimit-Limit`: Maximum requests allowed
    - `X-RateLimit-Remaining`: Requests remaining in current window
    - `X-RateLimit-Reset`: Time when the rate limit resets
    
    ## Error Handling
    
    The API uses standard HTTP status codes and returns error details in JSON format:
    
    ```json
    {
      "error": "Error message",
      "code": "ERROR_CODE",
      "details": {}
    }
    ```
  version: 1.0.0
  contact:
    name: FortiFi Team
    email: support@fortifi.dev
    url: https://fortifi.dev
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.fortifi.dev
    description: Production server
  - url: https://staging-api.fortifi.dev
...
paths:
...
  /api/token:
    delete:
      tags:
        - Authentication
      summary: Revoke token
      description: Revoke a JWT token to prevent further use
      operationId: revokeToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
                  description: JWT token to revoke
                  example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
              required:
                - token
      responses:
        '200':
          description: Token revoked successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    description: Whether the token was successfully revoked
              example:
                success: true
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/article/{id}:
    get:
      tags:
        - Articles
      summary: Get article content
      description: Retrieve article content. Requires valid JWT token for premium content.
      operationId: getArticle
      parameters:
        - name: id
          in: path
          required: true
          description: Article ID
          schema:
            type: string
          example: "article-123"
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Article content retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleResponse'
              example:
                id: "article-123"
                title: "The Future of Web Security"
                content: "<h2>Introduction</h2><p>Web security has evolved...</p>"
                metadata:
                  author: "Jane Smith"
                  publishedAt: "2024-01-15T10:00:00Z"
                  tags: ["security", "web", "authentication"]
                  category: "Security"
                assets:
                  pdf: "https://assets.fortifi.dev/articles/article-123.pdf"
                  images: ["https://assets.fortifi.dev/images/security.jpg"]
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Invalid or expired token"
                code: "INVALID_TOKEN"
        '403':
          description: Forbidden - Access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Access denied"
                code: "ACCESS_DENIED"
        '404':
          description: Article not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Article not found"
                code: "ARTICLE_NOT_FOUND"
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Rate limit exceeded"
                code: "RATE_LIMIT_EXCEEDED"
                details:
                  retryAfter: 60
    post:
      tags:
        - Articles
      summary: Create article
      description: Create a new article (admin only)
      operationId: createArticle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArticleRequest'
      responses:
        '201':
          description: Article created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - Articles
      summary: Update article
      description: Update an existing article (admin only)
      operationId: updateArticle
      parameters:
        - name: id
          in: path
          required: true
          description: Article ID
          schema:
            type: string
          example: "article-123"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArticleRequest'
      responses:
        '200':
          description: Article updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Article not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Articles
      summary: Delete article
      description: Delete an article (admin only)
      operationId: deleteArticle
      parameters:
        - name: id
          in: path
          required: true
          description: Article ID
          schema:
            type: string
          example: "article-123"
      responses:
        '200':
          description: Article deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    description: Whether the article was successfully deleted
              example:
                success: true
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Article not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
  schemas:
...
    ArticleRequest:
      type: object
      properties:
        title:
          type: string
          description: Article title
          example: "The Future of Web Security"
        content:
          type: string
          description: Article content (HTML)
          example: "<h2>Introduction</h2><p>Web security has evolved...</p>"
        metadata:
          type: object
          properties:
            author:
              type: string
              description: Article author
              example: "Jane Smith"
            publishedAt:
              type: string
              format: date-time
              description: Publication date
              example: "2024-01-15T10:00:00Z"
            tags:
              type: array
              items:
                type: string
              description: Article tags
              example: ["security", "web", "authentication"]
            category:
              type: string
              description: Article category
              example: "Security"
        assets:
          type: object
          properties:
            pdf:
              type: string
              format: uri
              description: PDF asset URL
              example: "https://assets.fortifi.dev/articles/article-123.pdf"
            images:
              type: array
              items:
                type: string
                format: uri
              description: Image asset URLs
              example: ["https://assets.fortifi.dev/images/security.jpg"]
      required:
        - title
        - content
        - metadata
    ArticleResponse:
      type: object
      properties:
        id:
          type: string
          description: Article ID
          example: "article-123"
        title:
          type: string
          description: Article title
          example: "The Future of Web Security"
        content:
          type: string
          description: Article content (HTML)
          example: "<h2>Introduction</h2><p>Web security has evolved...</p>"
        metadata:
          type: object
          properties:
            author:
              type: string
              description: Article author
              example: "Jane Smith"
            publishedAt:
              type: string
              format: date-time
              description: Publication date
              example: "2024-01-15T10:00:00Z"
            tags:
              type: array
              items:
                type: string
              description: Article tags
              example: ["security", "web", "authentication"]
            category:
              type: string
              description: Article category
              example: "Security"
          required:
            - author
            - publishedAt
            - tags
        assets:
          type: object
          properties:
            pdf:
              type: string
              format: uri
              description: PDF asset URL
              example: "https://assets.fortifi.dev/articles/article-123.pdf"
            images:
              type: array
              items:
                type: string
                format: uri
              description: Image asset URLs
              example: ["https://assets.fortifi.dev/images/security.jpg"]
      required:
        - id
        - title
        - content
        - metadata
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "Invalid or expired token"
        code:
          type: string
          description: Error code
          example: "INVALID_TOKEN"
        details:
          type: object
          description: Additional error details
          example:
            retryAfter: 60
      required:
        - error
...